chore: remove obsolete app folder
