--- src/app/(public)/acessos/page.tsx
@@ -1,2 +1,0 @@
-import Acesso from "@/pages/Acesso";
-export default Acesso;
