--- src/app/(public)/login/page.tsx
@@ -1,2 +1,0 @@
-import Login from "@/pages/Login";
-export default Login;
