--- src/app/(public)/reset/page.tsx
@@ -1,2 +1,0 @@
-import Reset from "@/pages/Reset";
-export default Reset;
