--- src/app/(public)/signup/page.tsx
@@ -1,2 +1,0 @@
-import Signup from "@/pages/Signup";
-export default Signup;
