--- src/app/globals.css
@@ -1,1 +1,0 @@
-@import '../index.css';
