+++ src/main.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from 'react-dom/client'
 import App from './App.tsx'
-import './app/globals.css'
+import './index.css'
 
 createRoot(document.getElementById("root")!).render(<App />);
