+++ tailwind.config.ts
@@ -4,10 +4,8 @@ import tailwindcssAnimate from "tailwindcss-animate";
 export default {
 	darkMode: ["class"],
 	content: [
-		"./pages/**/*.{ts,tsx}",
-		"./components/**/*.{ts,tsx}",
-		"./app/**/*.{ts,tsx}",
-		"./src/**/*.{ts,tsx}",
+                "./index.html",
+                "./src/**/*.{ts,tsx}",
 	],
 	prefix: "",
 	theme: {
